requirements.txt
streamlit
psutil
pyyaml
torch
transformers
sentencepiece
datasets
accelerate
